integrated nyan cat
