--- Nyan Cat Graphs/browserconfig.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Please read: https://msdn.microsoft.com/en-us/library/ie/dn455106.aspx -->
-<browserconfig>
-    <msapplication>
-        <tile>
-            <square70x70logo src="tile.png"/>
-            <square150x150logo src="tile.png"/>
-            <wide310x150logo src="tile-wide.png"/>
-            <square310x310logo src="tile.png"/>
-        </tile>
-    </msapplication>
-</browserconfig>
