--- Nyan Cat Graphs/robots.txt
@@ -1,5 +1,0 @@
-# www.robotstxt.org/
-
-# Allow crawling of all content
-User-agent: *
-Disallow:
