--- Nyan Cat Graphs/site.webmanifest
@@ -1,8 +1,0 @@
-{
-    "icons": [{
-        "src": "icon.png",
-        "sizes": "192x192",
-        "type": "image/png"
-    }],
-    "start_url": "/"
-}
