+++ src/app/players/components/player-nyan/player-nyan.component.html
@@ -1,5 +1,5 @@
 <div class="container">
-  <p id="wie">Wie gaat er om frieten?</p>
+  <p id="winner"></p>
   <div class="chartContainer">
     <canvas id="nyancat" #nyanCanvas>
     </canvas>
