+++ src/app/players/containers/player/player.component.ts
@@ -12,6 +12,14 @@ import { PlayerService } from '../../player.service';
 })
 export class PlayerComponent implements OnInit {
   players: Player[];
+  nyanify: boolean = false;
+  playing: boolean = false;
+
+
+  getDefaultPlayers(): void {
+    this.players = this.playerService.getDefaultPlayers();
+  }
+
 
   addPlayer(newPlayer): void {
     if (newPlayer) {
@@ -25,10 +33,34 @@ export class PlayerComponent implements OnInit {
     }
   }
 
+  removePlayer(i: number): void {
+    this.players.splice(i, 1);
+    this.playerService.updateLocalStorage(this.players);
+  }
+
+
+  startRandom(nyan): void {
+    // Shared
+    const playersLength: number = this.players.length;
+    this.playing = true;
+
+
+    if (nyan) {
+      // call nyan logic
+    } else {
+      const minScore = 5;
+      const maxScore: number = (playersLength >= (100 - minScore)) ? playersLength : 100;
+
+      this.playerService.assignRandomValues(this.players, playersLength, minScore, maxScore);
+    }
+
+  }
+
 
   constructor(private playerService: PlayerService) { }
 
   ngOnInit() {
+    this.getDefaultPlayers();
   }
 
 }
