+++ src/app/shop/components/product-list/product-list.component.html
@@ -1,4 +1,4 @@
-<div style="display: flex; justify-content: space-between">
+<div style="">
 
   <div class="card">
 
@@ -27,7 +27,7 @@
         {{ product.name }}
       </li>
     </ul>
-
+    <!--
     <form [formGroup]="basketForm">
 
       <input type="text" placeholder="your name" formControlName="name">
@@ -47,14 +47,15 @@
       </div>
 
       <button (click)="addAdditionalProduct()" class="btn btn-outline-primary no-margin-left">Add new Product</button>
-      <!--
+      !--
       <p>
         Value: {{ basketForm.value | json }}
       </p>
-      -->
+      --
       <br />
       <button type="submit" disabled class="btn btn-outline-primary no-margin-left">Submit</button>
     </form>
 
+  -->
   </div>
 </div>
