+++ src/styles.scss
@@ -147,10 +147,8 @@ input {
 }
 
 .close {
-    float: right;
     font-size: 1.5rem;
     font-weight: 700;
-    line-height: 1;
     color: #000;
     text-shadow: 0 1px 0 #fff;
     opacity: .5;
